add `close wallet` button to new tabbed settings page
