+++ app/components/views/SettingsPage/SettingsTab/Settings.module.css
@@ -214,18 +214,6 @@
   max-width: 400px;
 }
 
-.closeModalButton {
-  height: 23px;
-  background-color: var(--wallet-close-button-bg);
-  color: var(--wallet-close-button-text);
-  background-image: var(--close-wallet-icon);
-  background-repeat: no-repeat;
-  background-size: 13px;
-  background-position: 10px 8px;
-  border-radius: 10px;
-  padding: 6px 15px 0 30px;
-}
-
 .privacyButtons {
   text-align: right;
 }
