+++ app/components/views/SettingsPage/SettingsTab/SettingsTab.jsx
@@ -1,23 +1,9 @@
-import { FormattedMessage as T } from "react-intl";
 import ErrorScreen from "ErrorScreen";
 import SettingsPage from "./Settings";
-import { DescriptionHeader } from "layout";
 import { useSettings, useService } from "hooks";
 
-export const SettingsTabHeader = () => (
-  <DescriptionHeader
-    description={
-      <T
-        id="settings.description.settings"
-        m="Changing network settings requires a restart"
-      />
-    }
-  />
-);
-
 export const SettingsTab = () => {
-  const {
-    onCloseWallet,
+ const {
     tempSettings,
     onSaveSettings,
     onChangeTempSettings,
@@ -28,9 +14,7 @@ export const SettingsTab = () => {
     locales,
     networks,
     areSettingsDirty,
-    isTicketAutoBuyerEnabled,
     needNetworkReset,
-    walletName,
     walletReady
   } = useSettings();
   const { walletService } = useService();
@@ -47,16 +31,13 @@ export const SettingsTab = () => {
         onSaveSettings: onSaveSettingsHandler,
         changePassphraseRequestAttempt,
         isChangePassPhraseDisabled,
-        onCloseWallet,
         tempSettings,
         onChangeTempSettings,
         currencies,
         locales,
         networks,
         areSettingsDirty,
-        isTicketAutoBuyerEnabled,
         needNetworkReset,
-        walletName,
         walletReady
       }}
     />
